Transitioning between scenes now works fluidly.
Before, leaving the scene would result in either an abrupt main menu being shown which didn't feel nice for the player or the game would crash due to no reference for a coroutine.
Variables for setting the Fading In and Fading Out timer are also options available to the designer.