First level completed - Logan
The level is made up of three rooms.
Rooms are separated by moving platforms over death pits.
There are plenty of enemies. May need to tone it down after talking with Tom.
Had to have it load a previous level for the third room exit to not crash. Could have it load an empty room.
Level is probably harder due to being tested without other powers.