Reworked IEnvironment and added Enemy Death saving on checkpoints.
Restart was removed from all scripts using IEnvironment.
I recoded accordingly.

Doors now use Open and Close functions instead of the activate function which inverted the door's current status.
This was done in order to ensure that the door was in it's intended state and can never invert.

Entity spawners will now store a boolean that tracks whether or not their enemy has died.
In the scenario that the enemy dies and the player hits a checkpoint, the enemy will no longer respawn. This allows for longer areas to be made that allow backtracking among those areas.