Added Prefabs, Added deactivations to counters, implemented rechargeable batteries
Rechargeable Batteries
When the target is shot by electrokinesis, the battery is charged.
The battery will wait a duration before losing charge.
The battery will call its powered function when it has power and will call its noPower function when it runs out of power

Deactivation in Counters
Counters now have a boolean that can be toggled to allow the counter to deactivate whatever it's powering after initially powering it.

Prefabs
Added a Counter prefab to scene objects
Added a Battery prefab to scene objects